generalize helper functions
